Modified the entire pipeline to be population centric
Now the "sample" is the population, so the pipeline is not parallel per individual. Makes the node count dramatically smaller and easier to handle. Only the regime sampling has the potential to balloon the nodes as the other choices/variations will be kept inside of function/loops. This will largely likely be ok, INLA model might be the bottleneck in many ways but we will see.